Compiles in Ubuntu

+++ displaywidget.cpp
@@ -361,7 +361,7 @@ void DisplayWidget::
     // Find out what to crop based on selection
     unsigned FS = b->sample_rate();
     float radie = fabsf(selection[0].x - selection[1].x);
-    unsigned start = std::max(0.f, selection[0].x - radie/sqrt(2.f)) * FS;
+    unsigned start = std::max(0.f, selection[0].x - radie/sqrtf(2.f)) * FS;
     unsigned end = (selection[0].x + radie/sqrt(2.f)) * FS;
 
     if (end<=start)
